+++ lib/commands/clean.ts
@@ -19,7 +19,7 @@ export class CleanCommand implements ICommand {
 		const spinner = this.$terminalSpinnerService.createSpinner();
 		spinner.start("Cleaning project...\n");
 
-		const pathsToClean = [
+		let pathsToClean = [
 			constants.HOOKS_DIR_NAME,
 			constants.PLATFORMS_DIR_NAME,
 			constants.NODE_MODULES_FOLDER_NAME,
@@ -27,9 +27,18 @@ export class CleanCommand implements ICommand {
 		];
 
 		try {
-			const additionalPaths = this.$projectConfigService.getValue(
-				"additionalPathsToClean"
+			const overridePathsToClean = this.$projectConfigService.getValue(
+				"cli.pathsToClean"
 			);
+			const additionalPaths = this.$projectConfigService.getValue(
+				"cli.additionalPathsToClean"
+			);
+
+			// allow overriding default paths to clean
+			if (Array.isArray(overridePathsToClean)) {
+				pathsToClean = overridePathsToClean;
+			}
+
 			if (Array.isArray(additionalPaths)) {
 				pathsToClean.push(...additionalPaths);
 			}