+++ lib/commands/plugin/build-plugin.ts
@@ -52,6 +52,8 @@ export class BuildPluginCommand implements ICommand {
 		);
 
 		const options: IPluginBuildOptions = {
+			gradlePath: this.$options.gradlePath,
+			gradleArgs: this.$options.gradleArgs,
 			aarOutputDir: platformsAndroidPath,
 			platformsAndroidDirPath: platformsAndroidPath,
 			pluginName: pluginName,