+++ lib/controllers/prepare-controller.ts
@@ -18,6 +18,7 @@ import {
 	IProjectConfigService,
 	IProjectData,
 	IProjectDataService,
+	IProjectService,
 } from "../definitions/project";
 import {
 	INodeModulesDependenciesBuilder,
@@ -64,7 +65,8 @@ export class PrepareController extends EventEmitter {
 		private $watchIgnoreListService: IWatchIgnoreListService,
 		private $analyticsService: IAnalyticsService,
 		private $markingModeService: IMarkingModeService,
-		private $projectConfigService: IProjectConfigService
+		private $projectConfigService: IProjectConfigService,
+		private $projectService: IProjectService
 	) {
 		super();
 	}
@@ -127,7 +129,11 @@ export class PrepareController extends EventEmitter {
 		prepareData: IPrepareData,
 		projectData: IProjectData
 	): Promise<IPrepareResultData> {
-		await this.$platformController.addPlatformIfNeeded(prepareData);
+		await this.$projectService.ensureAppResourcesExist(projectData.projectDir);
+		await this.$platformController.addPlatformIfNeeded(
+			prepareData,
+			projectData
+		);
 		await this.trackRuntimeVersion(prepareData.platform, projectData);
 
 		this.$logger.info("Preparing project...");
@@ -375,7 +381,10 @@ export class PrepareController extends EventEmitter {
 		projectData: IProjectData
 	): Promise<string[]> {
 		const dependencies = this.$nodeModulesDependenciesBuilder
-			.getProductionDependencies(projectData.projectDir, projectData.ignoredDependencies)
+			.getProductionDependencies(
+				projectData.projectDir,
+				projectData.ignoredDependencies
+			)
 			.filter((dep) => dep.nativescript);
 		const pluginsNativeDirectories = dependencies.map((dep) =>
 			path.join(