+++ lib/options.ts
@@ -1,5 +1,6 @@
 import * as helpers from "./common/helpers";
 import * as yargs from "yargs";
+import { hideBin } from "yargs/helpers";
 import * as _ from "lodash";
 import {
 	IDictionary,
@@ -190,6 +191,7 @@ export class Options {
 			file: { type: OptionType.String, hasSensitiveValue: true },
 			force: { type: OptionType.Boolean, alias: "f", hasSensitiveValue: false },
 			emulator: { type: OptionType.Boolean, hasSensitiveValue: false },
+			simulator: { type: OptionType.Boolean, hasSensitiveValue: false },
 			sdk: { type: OptionType.String, hasSensitiveValue: false },
 			template: { type: OptionType.String, hasSensitiveValue: true },
 			certificate: { type: OptionType.String, hasSensitiveValue: true },
@@ -215,6 +217,8 @@ export class Options {
 				default: false,
 				hasSensitiveValue: false,
 			},
+			gradlePath: { type: OptionType.String, hasSensitiveValue: false },
+			gradleArgs: { type: OptionType.String, hasSensitiveValue: false },
 			aab: { type: OptionType.Boolean, hasSensitiveValue: false },
 			performance: { type: OptionType.Object, hasSensitiveValue: true },
 			appleApplicationSpecificPassword: {
@@ -364,9 +368,9 @@ export class Options {
 			opts[this.getDashedOptionName(key)] = value;
 		});
 
-		const parsed = yargs(process.argv.slice(2)).version(false).help(false);
-		this.initialArgv = parsed.argv;
-		this.argv = parsed.options(<any>opts).argv;
+		const parsed = yargs(hideBin(process.argv)).version(false).help(false);
+		this.initialArgv = parsed.argv as any;
+		this.argv = parsed.options(<any>opts).argv as any;
 
 		// For backwards compatibility
 		// Previously profileDir had a default option and calling `this.$options.profileDir` always returned valid result.
@@ -397,6 +401,11 @@ export class Options {
 
 		if (this.argv.javascript) {
 			this.argv.js = true;
+		}
+
+		// alias --simulator to --emulator
+		if (this.argv.simulator) {
+			this.argv.emulator = this.argv.simulator;
 		}
 
 		this.argv.bundle = "webpack";