+++ packages/doctor/src/android-tools-info.ts
@@ -29,12 +29,22 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 			"android-30",
 			"android-31",
 			"android-32",
+			"android-33",
 		];
 
-		if (runtimeVersion && semver.lt(semver.coerce(runtimeVersion), "6.1.0")) {
-			baseTargets.sort();
-			const indexOfSdk29 = baseTargets.indexOf("android-29");
-			baseTargets = baseTargets.slice(0, indexOfSdk29);
+		const isRuntimeVersionLessThan = (targetVersion: string) => {
+			return (
+				runtimeVersion &&
+				semver.lt(semver.coerce(runtimeVersion), targetVersion)
+			);
+		};
+
+		if (isRuntimeVersionLessThan("6.1.0")) {
+			// limit baseTargets to android-17 - android-28 if the runtime is < 6.1.0
+			baseTargets = baseTargets.slice(0, baseTargets.indexOf("android-29"));
+		} else if (isRuntimeVersionLessThan("8.2.0")) {
+			// limit baseTargets to android-17 - android-30 if the runtime is < 8.2.0
+			baseTargets = baseTargets.slice(0, baseTargets.indexOf("android-31"));
 		}
 
 		return baseTargets;
@@ -51,6 +61,8 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 		return process.env["ANDROID_HOME"];
 	}
 	private pathToEmulatorExecutable: string;
+
+	private _cachedRuntimeVersion: string;
 
 	constructor(
 		private childProcess: ChildProcess,
@@ -117,15 +129,16 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 				message = `You have to install version ${versionRangeMatches[1]}.`;
 			}
 
-			let invalidBuildToolsAdditionalMsg = `Run \`\$ ${this.getPathToSdkManagementTool()}\` from your command-line to install required \`Android Build Tools\`.`;
+			// let invalidBuildToolsAdditionalMsg = `Run \`\$ ${this.getPathToSdkManagementTool()}\` from your command-line to install required \`Android Build Tools\`.`;
+			let invalidBuildToolsAdditionalMsg = `Install the required build-tools through Android Studio.`;
 			if (!isAndroidHomeValid) {
 				invalidBuildToolsAdditionalMsg +=
-					" In case you already have them installed, make sure `ANDROID_HOME` environment variable is set correctly.";
+					" In case you already have them installed, make sure the `ANDROID_HOME` environment variable is set correctly.";
 			}
 
 			errors.push({
 				warning:
-					"You need to have the Android SDK Build-tools installed on your system. " +
+					"No compatible version of the Android SDK Build-tools are installed on your system. " +
 					message,
 				additionalInformation: invalidBuildToolsAdditionalMsg,
 				platforms: [Constants.ANDROID_PLATFORM_NAME],
@@ -502,13 +515,14 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 
 	private getMaxSupportedVersion(projectDir: string): number {
 		const supportedTargets = this.getSupportedTargets(projectDir);
+
 		return this.parseAndroidSdkString(
 			supportedTargets.sort()[supportedTargets.length - 1]
 		);
 	}
 
 	private getSystemRequirementsLink(): string {
-		return Constants.SYSTEM_REQUIREMENTS_LINKS[process.platform] || "";
+		return Constants.SYSTEM_REQUIREMENTS_LINKS;
 	}
 
 	private isAndroidHomeValid(): boolean {
@@ -598,6 +612,10 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 		runtimeVersion?: string;
 		projectDir?: string;
 	}): string {
+		if (this._cachedRuntimeVersion) {
+			return this._cachedRuntimeVersion;
+		}
+
 		let runtimePackage = {
 			name: Constants.ANDROID_SCOPED_RUNTIME,
 			version: runtimeVersion,
@@ -630,6 +648,7 @@ export class AndroidToolsInfo implements NativeScriptDoctor.IAndroidToolsInfo {
 			);
 		}
 
+		this._cachedRuntimeVersion = runtimeVersion;
 		return runtimeVersion;
 	}